Create a file for AudioStream and VideoStream. Reorganize code
