+++ src/net/RTP/MediaStream.cs
@@ -1,589 +1,12 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Net;
-using System.Net.Sockets;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using SIPSorcery.Net;
 using SIPSorcery.Sys;
-using SIPSorceryMedia.Abstractions;
 
 namespace SIPSorcery.net.RTP
 {
-
-    public class AudioStream : MediaStream
-    {
-        private static ILogger logger = Log.Logger;
-
-        /// <summary>
-        /// Gets fired when the remote SDP is received and the set of common audio formats is set.
-        /// </summary>
-        public event Action<List<AudioFormat>> OnAudioFormatsNegotiated;
-
-        public void CheckAudioFormatsNegotiation()
-        {
-            if(LocalTrack != null &&
-                        LocalTrack.Capabilities.Where(x => x.Name().ToLower() != SDP.TELEPHONE_EVENT_ATTRIBUTE).Count() > 0)
-            {
-                OnAudioFormatsNegotiated?.Invoke(
-                            LocalTrack.Capabilities
-                            .Where(x => x.Name().ToLower() != SDP.TELEPHONE_EVENT_ATTRIBUTE)
-                            .Select(x => x.ToAudioFormat()).ToList());
-            }
-        }
-
-        public AudioStream(RtpSessionConfig config) : base(config)
-        {
-            MediaType = SDPMediaTypesEnum.audio;
-        }
-
-        private Boolean rtpEventInProgress = false;
-
-        /// <summary>
-        /// Sends an audio sample to the remote peer.
-        /// </summary>
-        /// <param name="durationRtpUnits">The duration in RTP timestamp units of the audio sample. This
-        /// value is added to the previous RTP timestamp when building the RTP header.</param>
-        /// <param name="sample">The audio sample to set as the RTP packet payload.</param>
-        public void SendAudio(uint durationRtpUnits, byte[] sample)
-        {
-            if (DestinationEndPoint != null && ((!IsSecure && !UseSdpCryptoNegotiation) || IsSecurityContextReady()))
-            {
-                var audioFormat = GetSendingFormat();
-                SendAudioFrame(durationRtpUnits, audioFormat.ID, sample);
-            }
-        }
-
-        /// <summary>
-        /// Sends an audio packet to the remote party.
-        /// </summary>
-        /// <param name="duration">The duration of the audio payload in timestamp units. This value
-        /// gets added onto the timestamp being set in the RTP header.</param>
-        /// <param name="payloadTypeID">The payload ID to set in the RTP header.</param>
-        /// <param name="buffer">The audio payload to send.</param>
-        public void SendAudioFrame(uint duration, int payloadTypeID, byte[] buffer)
-        {
-            if (IsClosed || rtpEventInProgress || DestinationEndPoint == null || buffer == null || buffer.Length == 0)
-            {
-                return;
-            }
-
-            try
-            {
-                var audioTrack = LocalTrack;
-
-                if (audioTrack == null)
-                {
-                    logger.LogWarning("SendAudio was called on an RTP session without an audio stream.");
-                }
-                else if (!HasAudio || audioTrack.StreamStatus == MediaStreamStatusEnum.RecvOnly)
-                {
-                    return;
-                }
-                else
-                {
-                    // Basic RTP audio formats (such as G711, G722) do not have a concept of frames. The payload of the RTP packet is
-                    // considered a single frame. This results in a problem is the audio frame being sent is larger than the MTU. In 
-                    // that case the audio frame must be split across mutliple RTP packets. Unlike video frames theres no way to 
-                    // indicate that a series of RTP packets are correlated to the same timestamp. For that reason if an audio buffer
-                    // is supplied that's larger than MTU it will be split and the timestamp will be adjusted to best fit each RTP 
-                    // paylaod.
-                    // See https://github.com/sipsorcery/sipsorcery/issues/394.
-
-                    uint payloadTimestamp = audioTrack.Timestamp;
-                    uint payloadDuration = 0;
-
-                    for (int index = 0; index * RTPSession.RTP_MAX_PAYLOAD < buffer.Length; index++)
-                    {
-                        int offset = (index == 0) ? 0 : (index * RTPSession.RTP_MAX_PAYLOAD);
-                        int payloadLength = (offset + RTPSession.RTP_MAX_PAYLOAD < buffer.Length) ? RTPSession.RTP_MAX_PAYLOAD : buffer.Length - offset;
-                        payloadTimestamp += payloadDuration;
-                        byte[] payload = new byte[payloadLength];
-
-                        Buffer.BlockCopy(buffer, offset, payload, 0, payloadLength);
-
-                        // RFC3551 specifies that for audio the marker bit should always be 0 except for when returning
-                        // from silence suppression. For video the marker bit DOES get set to 1 for the last packet
-                        // in a frame.
-                        int markerBit = 0;
-
-                        var audioRtpChannel = rtpChannel;
-                        var protectRtpPacket = SecureContext?.ProtectRtpPacket;
-                        SendRtpPacket(audioRtpChannel, DestinationEndPoint, payload, payloadTimestamp, markerBit, payloadTypeID, audioTrack.Ssrc, audioTrack.GetNextSeqNum(), RtcpSession, protectRtpPacket);
-
-                        //logger.LogDebug($"send audio { audioRtpChannel.RTPLocalEndPoint}->{AudioDestinationEndPoint}.");
-
-                        payloadDuration = (uint)(((decimal)payloadLength / buffer.Length) * duration); // Get the percentage duration of this payload.
-                    }
-
-                    audioTrack.Timestamp += duration;
-                }
-            }
-            catch (SocketException sockExcp)
-            {
-                logger.LogError("SocketException SendAudioFrame. " + sockExcp.Message);
-            }
-        }
-
-        /// <summary>
-        /// Sends an RTP event for a DTMF tone as per RFC2833. Sending the event requires multiple packets to be sent.
-        /// This method will hold onto the socket until all the packets required for the event have been sent. The send
-        /// can be cancelled using the cancellation token.
-        /// </summary>
-        /// <param name="rtpEvent">The RTP event to send.</param>
-        /// <param name="cancellationToken">CancellationToken to allow the operation to be cancelled prematurely.</param>
-        /// <param name="clockRate">To send an RTP event the clock rate of the underlying stream needs to be known.</param>
-        /// <param name="samplePeriod">The sample period in milliseconds being used for the media stream that the event 
-        /// is being inserted into. Should be set to 50ms if main media stream is dynamic or sample period is unknown.</param>
-        public async Task SendDtmfEvent(RTPEvent rtpEvent, CancellationToken cancellationToken, int clockRate = RTPSession.DEFAULT_AUDIO_CLOCK_RATE, int samplePeriod = RTPSession.RTP_EVENT_DEFAULT_SAMPLE_PERIOD_MS)
-        {
-            if (IsClosed || rtpEventInProgress == true || DestinationEndPoint == null)
-            {
-                logger.LogWarning("SendDtmfEvent request ignored as an RTP event is already in progress.");
-            }
-
-            try
-            {
-                if (LocalTrack == null)
-                {
-                    logger.LogWarning("SendDtmfEvent was called on an RTP session without an audio stream.");
-                }
-                else if (!HasAudio || LocalTrack.StreamStatus == MediaStreamStatusEnum.RecvOnly)
-                {
-                    return;
-                }
-                else
-                {
-                    rtpEventInProgress = true;
-                    uint startTimestamp = LocalTrack.Timestamp;
-
-                    // The RTP timestamp step corresponding to the sampling period. This can change depending
-                    // on the codec being used. For example using PCMU with a sampling frequency of 8000Hz and a sample period of 50ms
-                    // the timestamp step is 400 (8000 / (1000 / 50)). For a sample period of 20ms it's 160 (8000 / (1000 / 20)).
-                    ushort rtpTimestampStep = (ushort)(clockRate * samplePeriod / 1000);
-
-                    // If only the minimum number of packets are being sent then they are both the start and end of the event.
-                    rtpEvent.EndOfEvent = (rtpEvent.TotalDuration <= rtpTimestampStep);
-                    // The DTMF tone is generally multiple RTP events. Each event has a duration of the RTP timestamp step.
-                    rtpEvent.Duration = rtpTimestampStep;
-
-                    // Send the start of event packets.
-                    for (int i = 0; i < RTPEvent.DUPLICATE_COUNT && !cancellationToken.IsCancellationRequested; i++)
-                    {
-                        byte[] buffer = rtpEvent.GetEventPayload();
-
-                        int markerBit = (i == 0) ? 1 : 0;  // Set marker bit for the first packet in the event.
-                        var protectRtpPacket = SecureContext?.ProtectRtpPacket;
-                        SendRtpPacket(rtpChannel, DestinationEndPoint, buffer, startTimestamp, markerBit, rtpEvent.PayloadTypeID, LocalTrack.Ssrc, LocalTrack.GetNextSeqNum(), RtcpSession, protectRtpPacket);
-                    }
-
-                    await Task.Delay(samplePeriod, cancellationToken).ConfigureAwait(false);
-
-                    if (!rtpEvent.EndOfEvent)
-                    {
-                        // Send the progressive event packets 
-                        while ((rtpEvent.Duration + rtpTimestampStep) < rtpEvent.TotalDuration && !cancellationToken.IsCancellationRequested)
-                        {
-                            rtpEvent.Duration += rtpTimestampStep;
-                            byte[] buffer = rtpEvent.GetEventPayload();
-
-                            var protectRtpPacket = SecureContext?.ProtectRtpPacket;
-                            SendRtpPacket(rtpChannel, DestinationEndPoint, buffer, startTimestamp, 0, rtpEvent.PayloadTypeID, LocalTrack.Ssrc, LocalTrack.GetNextSeqNum(), RtcpSession, protectRtpPacket);
-
-                            await Task.Delay(samplePeriod, cancellationToken).ConfigureAwait(false);
-                        }
-
-                        // Send the end of event packets.
-                        for (int j = 0; j < RTPEvent.DUPLICATE_COUNT && !cancellationToken.IsCancellationRequested; j++)
-                        {
-                            rtpEvent.EndOfEvent = true;
-                            rtpEvent.Duration = rtpEvent.TotalDuration;
-                            byte[] buffer = rtpEvent.GetEventPayload();
-
-                            var protectRtpPacket = SecureContext?.ProtectRtpPacket;
-                            SendRtpPacket(rtpChannel, DestinationEndPoint, buffer, startTimestamp, 0, rtpEvent.PayloadTypeID, LocalTrack.Ssrc, LocalTrack.GetNextSeqNum(), RtcpSession, protectRtpPacket);
-                        }
-                    }
-                    LocalTrack.Timestamp += rtpEvent.TotalDuration;
-                }
-            }
-            catch (SocketException sockExcp)
-            {
-                logger.LogError("SocketException SendDtmfEvent. " + sockExcp.Message);
-            }
-            catch (TaskCanceledException)
-            {
-                logger.LogWarning("SendDtmfEvent was cancelled by caller.");
-            }
-            finally
-            {
-                rtpEventInProgress = false;
-            }
-        }
-
-        /// <summary>
-        /// Sends a DTMF tone as an RTP event to the remote party.
-        /// </summary>
-        /// <param name="key">The DTMF tone to send.</param>
-        /// <param name="ct">RTP events can span multiple RTP packets. This token can
-        /// be used to cancel the send.</param>
-        public virtual Task SendDtmf(byte key, CancellationToken ct)
-        {
-            var dtmfEvent = new RTPEvent(key, false, RTPEvent.DEFAULT_VOLUME, RTPSession.DTMF_EVENT_DURATION, RTPSession.DTMF_EVENT_PAYLOAD_ID);
-            return SendDtmfEvent(dtmfEvent, ct);
-        }
-
-        /// <summary>
-        /// Indicates whether this session is using audio.
-        /// </summary>
-        public bool HasAudio
-        {
-            get
-            {
-                return LocalTrack != null && LocalTrack.StreamStatus != MediaStreamStatusEnum.Inactive
-                  && RemoteTrack != null && RemoteTrack.StreamStatus != MediaStreamStatusEnum.Inactive;
-            }
-        }
-
-        public void OnReceiveRTPPacket(RTPHeader hdr, SDPAudioVideoMediaFormat format, int localPort, IPEndPoint remoteEndPoint, byte[] buffer)
-        {
-            base.OnReceiveRTPPacket(hdr, format, localPort, remoteEndPoint, buffer, null);
-        }
-    }
-
-    public class VideoStream: MediaStream
-    {
-        private static ILogger logger = Log.Logger;
-
-        /// <summary>
-        /// Gets fired when the remote SDP is received and the set of common video formats is set.
-        /// </summary>
-        public event Action<List<VideoFormat>> OnVideoFormatsNegotiated;
-
-        /// <summary>
-        /// Gets fired when a full video frame is reconstructed from one or more RTP packets
-        /// received from the remote party.
-        /// </summary>
-        /// <remarks>
-        ///  - Received from end point,
-        ///  - The frame timestamp,
-        ///  - The encoded video frame payload.
-        ///  - The video format of the encoded frame.
-        /// </remarks>
-        public event Action<IPEndPoint, uint, byte[], VideoFormat> OnVideoFrameReceived;
-
-        public RtpVideoFramer RtpVideoFramer;
-
-        /// <summary>
-        /// Indicates whether this session is using video.
-        /// </summary>
-        public bool HasVideo
-        {
-            get
-            {
-                return LocalTrack != null && LocalTrack.StreamStatus != MediaStreamStatusEnum.Inactive
-                  && RemoteTrack != null && RemoteTrack.StreamStatus != MediaStreamStatusEnum.Inactive;
-            }
-        }
-
-        public void CheckVideoFormatsNegotiation()
-        {
-            if (LocalTrack != null && LocalTrack.Capabilities?.Count() > 0)
-            {
-                OnVideoFormatsNegotiated?.Invoke(
-                            LocalTrack.Capabilities
-                            .Select(x => x.ToVideoFormat()).ToList());
-            }
-        }
-
-        public void ProcessVideoRtpFrame(IPEndPoint endpoint, RTPPacket packet, SDPAudioVideoMediaFormat format)
-        {
-            if (OnVideoFrameReceived == null)
-            {
-                return;
-            }
-
-            if (RtpVideoFramer != null)
-            {
-                var frame = RtpVideoFramer.GotRtpPacket(packet);
-                if (frame != null)
-                {
-                    OnVideoFrameReceived?.Invoke(endpoint, packet.Header.Timestamp, frame, format.ToVideoFormat());
-                }
-            }
-            else
-            {
-                if (format.ToVideoFormat().Codec == VideoCodecsEnum.VP8 ||
-                    format.ToVideoFormat().Codec == VideoCodecsEnum.H264)
-                {
-                    logger.LogDebug($"Video depacketisation codec set to {format.ToVideoFormat().Codec} for SSRC {packet.Header.SyncSource}.");
-
-                    RtpVideoFramer = new RtpVideoFramer(format.ToVideoFormat().Codec);
-
-                    var frame = RtpVideoFramer.GotRtpPacket(packet);
-                    if (frame != null)
-                    {
-                        OnVideoFrameReceived?.Invoke(endpoint, packet.Header.Timestamp, frame, format.ToVideoFormat());
-                    }
-                }
-                else
-                {
-                    logger.LogWarning($"Video depacketisation logic for codec {format.Name()} has not been implemented, PR's welcome!");
-                }
-            }
-        }
-
-        /// <summary>
-        /// Helper method to send a low quality JPEG image over RTP. This method supports a very abbreviated version of RFC 2435 "RTP Payload Format for JPEG-compressed Video".
-        /// It's intended as a quick convenient way to send something like a test pattern image over an RTSP connection. More than likely it won't be suitable when a high
-        /// quality image is required since the header used in this method does not support quantization tables.
-        /// </summary>
-        /// <param name="jpegBytes">The raw encoded bytes of the JPEG image to transmit.</param>
-        /// <param name="jpegQuality">The encoder quality of the JPEG image.</param>
-        /// <param name="jpegWidth">The width of the JPEG image.</param>
-        /// <param name="jpegHeight">The height of the JPEG image.</param>
-        /// <param name="framesPerSecond">The rate at which the JPEG frames are being transmitted at. used to calculate the timestamp.</param>
-        public void SendJpegFrame(uint duration, int payloadTypeID, byte[] jpegBytes, int jpegQuality, int jpegWidth, int jpegHeight)
-        {
-            if (IsClosed || DestinationEndPoint == null)
-            {
-                return;
-            }
-            try
-            {
-                var videoTrack = LocalTrack;
-
-                if (videoTrack == null)
-                {
-                    logger.LogWarning("SendJpegFrame was called on an RTP session without a video stream.");
-                }
-                else if (!HasVideo || videoTrack.StreamStatus == MediaStreamStatusEnum.RecvOnly)
-                {
-                    return;
-                }
-                else
-                {
-                    for (int index = 0; index * RTPSession.RTP_MAX_PAYLOAD < jpegBytes.Length; index++)
-                    {
-                        uint offset = Convert.ToUInt32(index * RTPSession.RTP_MAX_PAYLOAD);
-                        int payloadLength = ((index + 1) * RTPSession.RTP_MAX_PAYLOAD < jpegBytes.Length) ? RTPSession.RTP_MAX_PAYLOAD : jpegBytes.Length - index * RTPSession.RTP_MAX_PAYLOAD;
-                        byte[] jpegHeader = RtpVideoFramer.CreateLowQualityRtpJpegHeader(offset, jpegQuality, jpegWidth, jpegHeight);
-
-                        List<byte> packetPayload = new List<byte>();
-                        packetPayload.AddRange(jpegHeader);
-                        packetPayload.AddRange(jpegBytes.Skip(index * RTPSession.RTP_MAX_PAYLOAD).Take(payloadLength));
-
-                        int markerBit = ((index + 1) * RTPSession.RTP_MAX_PAYLOAD < jpegBytes.Length) ? 0 : 1;
-
-                        var protectRtpPacket = SecureContext?.ProtectRtpPacket;
-                        SendRtpPacket(rtpChannel, DestinationEndPoint, packetPayload.ToArray(), videoTrack.Timestamp, markerBit, payloadTypeID, videoTrack.Ssrc, videoTrack.GetNextSeqNum(), RtcpSession, protectRtpPacket);
-                    }
-
-                    videoTrack.Timestamp += duration;
-                }
-            }
-            catch (SocketException sockExcp)
-            {
-                logger.LogError("SocketException SendJpegFrame. " + sockExcp.Message);
-            }
-        }
-
-        /// <summary>
-        /// Sends a H264 frame, represented by an Access Unit, to the remote party.
-        /// </summary>
-        /// <param name="duration">The duration in timestamp units of the payload (e.g. 3000 for 30fps).</param>
-        /// <param name="payloadTypeID">The payload type ID  being used for H264 and that will be set on the RTP header.</param>
-        /// <param name="accessUnit">The encoded H264 access unit to transmit. An access unit can contain one or more
-        /// NAL's.</param>
-        /// <remarks>
-        /// An Access Unit can contain one or more NAL's. The NAL's have to be parsed in order to be able to package 
-        /// in RTP packets.
-        /// 
-        /// See https://www.itu.int/rec/dologin_pub.asp?lang=e&id=T-REC-H.264-201602-S!!PDF-E&type=items Annex B for byte stream specification.
-        /// </remarks>
-        public void SendH264Frame(uint duration, int payloadTypeID, byte[] accessUnit)
-        {
-            var dstEndPoint = DestinationEndPoint;
-
-            if (IsClosed || dstEndPoint == null || accessUnit == null || accessUnit.Length == 0)
-            {
-                return;
-            }
-
-            if (LocalTrack == null)
-            {
-                logger.LogWarning("SendH264Frame was called on an RTP session without a video stream.");
-            }
-            else if (!HasVideo || LocalTrack.StreamStatus == MediaStreamStatusEnum.RecvOnly)
-            {
-                return;
-            }
-            else
-            {
-                foreach (var nal in H264Packetiser.ParseNals(accessUnit))
-                {
-                    SendH264Nal(duration, payloadTypeID, nal.NAL, nal.IsLast, dstEndPoint, LocalTrack);
-                }
-            }
-        }
-
-        /// <summary>
-        /// Sends a single H264 NAL to the remote party.
-        /// </summary>
-        /// <param name="duration">The duration in timestamp units of the payload (e.g. 3000 for 30fps).</param>
-        /// <param name="payloadTypeID">The payload type ID  being used for H264 and that will be set on the RTP header.</param>
-        /// <param name="nal">The buffer containing the NAL to send.</param>
-        /// <param name="isLastNal">Should be set for the last NAL in the H264 access unit. Determines when the markbit gets set 
-        /// and the timestamp incremented.</param>
-        /// <param name="dstEndPoint">The destination end point to send to.</param>
-        /// <param name="videoTrack">The video track to send on.</param>
-        private void SendH264Nal(uint duration, int payloadTypeID, byte[] nal, bool isLastNal, IPEndPoint dstEndPoint, MediaStreamTrack videoTrack)
-        {
-            //logger.LogDebug($"Send NAL {nal.Length}, is last {isLastNal}, timestamp {videoTrack.Timestamp}.");
-            //logger.LogDebug($"nri {nalNri:X2}, type {nalType:X2}.");
-
-            byte nal0 = nal[0];
-
-            if (nal.Length <= RTPSession.RTP_MAX_PAYLOAD)
-            {
-                // Send as Single-Time Aggregation Packet (STAP-A).
-                byte[] payload = new byte[nal.Length];
-                int markerBit = isLastNal ? 1 : 0;   // There is only ever one packet in a STAP-A.
-                Buffer.BlockCopy(nal, 0, payload, 0, nal.Length);
-
-                var protectRtpPacket = SecureContext?.ProtectRtpPacket;
-                SendRtpPacket(rtpChannel, dstEndPoint, payload, videoTrack.Timestamp, markerBit, payloadTypeID, videoTrack.Ssrc, videoTrack.GetNextSeqNum(), RtcpSession, protectRtpPacket);
-                //logger.LogDebug($"send H264 {videoChannel.RTPLocalEndPoint}->{dstEndPoint} timestamp {videoTrack.Timestamp}, payload length {payload.Length}, seqnum {videoTrack.SeqNum}, marker {markerBit}.");
-                //logger.LogDebug($"send H264 {videoChannel.RTPLocalEndPoint}->{dstEndPoint} timestamp {videoTrack.Timestamp}, STAP-A {h264RtpHdr.HexStr()}, payload length {payload.Length}, seqnum {videoTrack.SeqNum}, marker {markerBit}.");
-            }
-            else
-            {
-                nal = nal.Skip(1).ToArray();
-
-                // Send as Fragmentation Unit A (FU-A):
-                for (int index = 0; index * RTPSession.RTP_MAX_PAYLOAD < nal.Length; index++)
-                {
-                    int offset = index * RTPSession.RTP_MAX_PAYLOAD;
-                    int payloadLength = ((index + 1) * RTPSession.RTP_MAX_PAYLOAD < nal.Length) ? RTPSession.RTP_MAX_PAYLOAD : nal.Length - index * RTPSession.RTP_MAX_PAYLOAD;
-
-                    bool isFirstPacket = index == 0;
-                    bool isFinalPacket = (index + 1) * RTPSession.RTP_MAX_PAYLOAD >= nal.Length;
-                    int markerBit = (isLastNal && isFinalPacket) ? 1 : 0;
-
-                    byte[] h264RtpHdr = H264Packetiser.GetH264RtpHeader(nal0, isFirstPacket, isFinalPacket);
-
-                    byte[] payload = new byte[payloadLength + h264RtpHdr.Length];
-                    Buffer.BlockCopy(h264RtpHdr, 0, payload, 0, h264RtpHdr.Length);
-                    Buffer.BlockCopy(nal, offset, payload, h264RtpHdr.Length, payloadLength);
-
-
-                    var protectRtpPacket = SecureContext?.ProtectRtpPacket;
-                    SendRtpPacket(rtpChannel, dstEndPoint, payload, videoTrack.Timestamp, markerBit, payloadTypeID, videoTrack.Ssrc, videoTrack.GetNextSeqNum(), RtcpSession, protectRtpPacket);
-                    //logger.LogDebug($"send H264 {videoChannel.RTPLocalEndPoint}->{dstEndPoint} timestamp {videoTrack.Timestamp}, FU-A {h264RtpHdr.HexStr()}, payload length {payloadLength}, seqnum {videoTrack.SeqNum}, marker {markerBit}.");
-                }
-            }
-
-            if (isLastNal)
-            {
-                videoTrack.Timestamp += duration;
-            }
-        }
-
-        /// <summary>
-        /// Sends a VP8 frame as one or more RTP packets.
-        /// </summary>
-        /// <param name="timestamp">The timestamp to place in the RTP header. Needs
-        /// to be based on a 90Khz clock.</param>
-        /// <param name="payloadTypeID">The payload ID to place in the RTP header.</param>
-        /// <param name="buffer">The VP8 encoded payload.</param>
-        public void SendVp8Frame(uint duration, int payloadTypeID, byte[] buffer)
-        {
-            if (IsClosed || DestinationEndPoint == null)
-            {
-                return;
-            }
-
-            try
-            {
-                if (LocalTrack == null)
-                {
-                    logger.LogWarning("SendVp8Frame was called on an RTP session without a video stream.");
-                }
-                else if (!HasVideo || LocalTrack.StreamStatus == MediaStreamStatusEnum.RecvOnly)
-                {
-                    return;
-                }
-                else
-                {
-                    for (int index = 0; index * RTPSession.RTP_MAX_PAYLOAD < buffer.Length; index++)
-                    {
-                        int offset = index * RTPSession.RTP_MAX_PAYLOAD;
-                        int payloadLength = (offset + RTPSession.RTP_MAX_PAYLOAD < buffer.Length) ? RTPSession.RTP_MAX_PAYLOAD : buffer.Length - offset;
-
-                        byte[] vp8HeaderBytes = (index == 0) ? new byte[] { 0x10 } : new byte[] { 0x00 };
-                        byte[] payload = new byte[payloadLength + vp8HeaderBytes.Length];
-                        Buffer.BlockCopy(vp8HeaderBytes, 0, payload, 0, vp8HeaderBytes.Length);
-                        Buffer.BlockCopy(buffer, offset, payload, vp8HeaderBytes.Length, payloadLength);
-
-                        int markerBit = ((offset + payloadLength) >= buffer.Length) ? 1 : 0; // Set marker bit for the last packet in the frame.
-
-                        var protectRtpPacket = SecureContext?.ProtectRtpPacket;
-                        SendRtpPacket(rtpChannel, DestinationEndPoint, payload, LocalTrack.Timestamp, markerBit, payloadTypeID, LocalTrack.Ssrc, LocalTrack.GetNextSeqNum(), RtcpSession, protectRtpPacket);
-                        //logger.LogDebug($"send VP8 {videoChannel.RTPLocalEndPoint}->{dstEndPoint} timestamp {videoTrack.Timestamp}, sample length {buffer.Length}.");
-                    }
-
-                    LocalTrack.Timestamp += duration;
-                }
-            }
-            catch (SocketException sockExcp)
-            {
-                logger.LogError("SocketException SendVp8Frame. " + sockExcp.Message);
-            }
-        }
-
-        /// <summary>
-        /// Sends a video sample to the remote peer.
-        /// </summary>
-        /// <param name="durationRtpUnits">The duration in RTP timestamp units of the video sample. This
-        /// value is added to the previous RTP timestamp when building the RTP header.</param>
-        /// <param name="sample">The video sample to set as the RTP packet payload.</param>
-        public void SendVideo(uint durationRtpUnits, byte[] sample)
-        {
-            if (DestinationEndPoint != null || (DestinationEndPoint != null) &&
-                ((!IsSecure && !UseSdpCryptoNegotiation) || IsSecurityContextReady()))
-            {
-                var videoSendingFormat = GetSendingFormat();
-
-                switch (videoSendingFormat.Name())
-                {
-                    case "VP8":
-                        int vp8PayloadID = Convert.ToInt32(LocalTrack.Capabilities.First(x => x.Name() == "VP8").ID);
-                        SendVp8Frame(durationRtpUnits, vp8PayloadID, sample);
-                        break;
-                    case "H264":
-                        int h264PayloadID = Convert.ToInt32(LocalTrack.Capabilities.First(x => x.Name() == "H264").ID);
-                        SendH264Frame(durationRtpUnits, h264PayloadID, sample);
-                        break;
-                    default:
-                        throw new ApplicationException($"Unsupported video format selected {videoSendingFormat.Name()}.");
-                }
-            }
-        }
-
-        public void OnReceiveRTPPacket(RTPHeader hdr, SDPAudioVideoMediaFormat format, int localPort, IPEndPoint remoteEndPoint, byte[] buffer)
-        {
-            base.OnReceiveRTPPacket(hdr, format, localPort, remoteEndPoint, buffer, this);
-        }
-
-        public VideoStream(RtpSessionConfig config) : base(config)
-        {
-            MediaType = SDPMediaTypesEnum.video;
-        }
-    }
-
     public class MediaStream
     {
         private static ILogger logger = Log.Logger;
@@ -599,7 +22,7 @@ namespace SIPSorcery.net.RTP
 
         private RTPReorderBuffer RTPReorderBuffer = null;
 
-        public Boolean AcceptRtpFromAny { get; set; } = false;
+        protected RTPChannel rtpChannel = null;
 
     #region EVENTS
 
@@ -613,7 +36,6 @@ namespace SIPSorcery.net.RTP
         /// Gets fired when an RTCP report is sent. This event is for diagnostics only.
         /// </summary>
         public event Action<SDPMediaTypesEnum, RTCPCompoundPacket> OnSendReport;
-
 
         /// <summary>
         /// Gets fired when an RTP packet is received from a remote party.
@@ -638,7 +60,7 @@ namespace SIPSorcery.net.RTP
 
     #region PROPERTIES
 
-        protected RTPChannel rtpChannel = null;
+        public Boolean AcceptRtpFromAny { get; set; } = false;
 
         /// <summary>
         /// Indicates whether the session has been closed. Once a session is closed it cannot
@@ -946,6 +368,8 @@ namespace SIPSorcery.net.RTP
 
     #endregion SEND PACKET
 
+    #region RECEIVE PACKET
+
         protected void OnReceiveRTPPacket(RTPHeader hdr, SDPAudioVideoMediaFormat format, int localPort, IPEndPoint remoteEndPoint, byte[] buffer, VideoStream videoStream = null)
         {
             RTPPacket rtpPacket = null;
@@ -1026,6 +450,10 @@ namespace SIPSorcery.net.RTP
             }
         }
 
+    #endregion RECEIVE PACKET
+
+    #region TO RAISE EVENTS FROM INHERITED CLASS
+
         public void ReceiveReport(IPEndPoint ipEndPoint, RTCPCompoundPacket rtcpPCompoundPacket)
         {
             OnReceiveReport?.Invoke(ipEndPoint, MediaType, rtcpPCompoundPacket);
@@ -1040,6 +468,8 @@ namespace SIPSorcery.net.RTP
         {
             OnRtpPacketReceived?.Invoke(ipEndPoint, MediaType, rtpPacket);
         }
+
+    #endregion TO RAISE EVENTS FROM INHERITED CLASS
 
         protected void LogIfWrongSeqNumber(string trackType, RTPHeader header, MediaStreamTrack track)
         {
@@ -1104,13 +534,6 @@ namespace SIPSorcery.net.RTP
             }
 
             return isValidSource;
-        }
-
-        public MediaStream(RtpSessionConfig config)
-        {
-            IsSecure = config.RtpSecureMediaOption == RtpSecureMediaOptionEnum.DtlsSrtp;
-            UseSdpCryptoNegotiation = config.RtpSecureMediaOption == RtpSecureMediaOptionEnum.SdpCryptoNegotiation;
-            RtpSessionConfig = config;
         }
 
         /// <summary>
@@ -1200,5 +623,13 @@ namespace SIPSorcery.net.RTP
                 }
             });
         }
+
+        public MediaStream(RtpSessionConfig config)
+        {
+            IsSecure = config.RtpSecureMediaOption == RtpSecureMediaOptionEnum.DtlsSrtp;
+            UseSdpCryptoNegotiation = config.RtpSecureMediaOption == RtpSecureMediaOptionEnum.SdpCryptoNegotiation;
+            RtpSessionConfig = config;
+        }
     }
+
 }
